Conflicts:
	README.md

+++ test/test_column_basic.cpp
@@ -15,8 +15,8 @@ TEST(ColumnBasic_LowerUpperBound)
     // Create column with sorted members
     BasicColumn<int> col;
     col.add(5);
-    for (size_t i = 5; i < 100; i += 5) {
-        col.add(int(i));
+    for (int i = 5; i < 100; i += 5) {
+        col.add(i);
     }
 
     // before first entry
